+++ etebase_server/django/utils.py
