+++ etebase_server/fastapi/routers/invitation.py
@@ -4,8 +4,8 @@ from django.db import transaction, IntegrityError
 from django.db.models import QuerySet
 from fastapi import APIRouter, Depends, status, Request
 
-from django_etebase import models
-from django_etebase.utils import get_user_queryset, CallbackContext
+from etebase_server.django import models
+from etebase_server.django.utils import get_user_queryset, CallbackContext
 from etebase_server.myauth.models import UserType, get_typed_user_model
 from .authentication import get_authenticated_user
 from ..exceptions import HttpError, PermissionDenied
